% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_cell_ratio using this to avoid function name confiction.}
\alias{add_cell_ratio using this to avoid function name confiction.}
\alias{add_cell_ratio}
\title{add_cell_ratio}
\usage{
add_cell_ratio(sce, celltypes = "cell_type10")
}
\arguments{
\item{sce}{object, a singcellexperiment object.}
}
\value{
a singcellexperiment object
}
\description{
add_cell_ratio
}
