% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2d.R
\name{boxplot_clinical}
\alias{boxplot_clinical}
\title{Visualization of cluster data of sce}
\usage{
boxplot_clinical(
  df = df,
  sel_group = "stage",
  celltype = "DC",
  axistheme = axistheme1,
  compare = compare_list
)
}
\arguments{
\item{df}{An sce object}

\item{sel_group}{character. Items use to 2D plot, axes x and y must be numeric.}

\item{celltype}{character. Dot or mesh color by which character. It can be one of the column
of plot_meta, or it can be just "density" (the default value).}
}
\value{
ggplot2 figure
}
\description{
Visualization of cluster data of sce
}
