% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{cal_closest_dist}
\alias{cal_closest_dist}
\title{cal_closest_dist}
\usage{
cal_closest_dist(roi, cld, query_cells, taget_cells, k = 1)
}
\arguments{
\item{roi}{object, a singcellexperiment object}

\item{cld}{object, a singcellexperiment object}

\item{query_cells}{object, a singcellexperiment object}

\item{taget_cells}{object, a singcellexperiment object}

\item{k}{object, a singcellexperiment object}
}
\description{
cal_closest_dist
}
