% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_distance.R
\name{cal_distance using this to avoid function name confiction.}
\alias{cal_distance using this to avoid function name confiction.}
\alias{cal_distance}
\alias{cal_distance_all}
\title{cal_distance}
\usage{
cal_distance(sce, tocell = "CD8T")

cal_distance_all(sce)
}
\arguments{
\item{sce}{object, a singcellexperiment object.}

\item{tocell}{string, CD8T.}
}
\value{
A data frame.

A data frame.
}
\description{
cal_distance

cal_distance
}
\examples{
\dontrun{
sce <- cal_distance(sce, tocell = "CD8T")
df <- cal_distance(sce)
}
\dontrun{
sce <- cal_distance(sce, tocell = "CD8T")
a <- cal_distance_all(sce)
}
}
