% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytomapper_plot.R
\name{cytomapper_cell_plot}
\alias{cytomapper_cell_plot}
\title{cytomapper_cell_plot}
\usage{
cytomapper_cell_plot(
  sce,
  sample_id,
  dat_dir,
  colour_by = "cluster",
  pdf_fn,
  panel_fn = "~/projects/hyperion/data/qzhang/human/hyperion/panel.csv"
)
}
\arguments{
\item{colour_by}{colour_by = c("Collagen1", "Vimentin", "DNA1", "E_Cadherin", "a_smooth")}

\item{bcg}{bcg = list(Collagen1 = c(0, 20, 1), Vimentin = c(0, 20, 1), DNA1 = c(0, 2, 1), E_Cadherin = c(0, 20, 1), a_smooth = c(0, 20, 1))}
}
\description{
cytomapper_cell_plot
}
