% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytomapper_plot.R
\name{cytomapper_plot}
\alias{cytomapper_plot}
\title{cytomapper_plot}
\usage{
cytomapper_plot(
  dat_dir,
  pdf_fn,
  sample_id,
  colour_by = c("CD3", "CD20"),
  bcg = list(CD3 = c(0, 20, 1), CD20 = c(0, 20, 1)),
  display = "single",
  panel_fn = "~/projects/hyperion/data/qzhang/human/hyperion/panel.csv",
  image_title = NULL
)
}
\arguments{
\item{colour_by}{colour_by = c("Collagen1", "Vimentin", "DNA1", "E_Cadherin", "a_smooth")}

\item{bcg}{bcg = list(Collagen1 = c(0, 20, 1), Vimentin = c(0, 20, 1), DNA1 = c(0, 2, 1), E_Cadherin = c(0, 20, 1), a_smooth = c(0, 20, 1))}
}
\description{
cytomapper_plot
}
