% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_distance.R
\name{dist_min using this to avoid function name confiction.}
\alias{dist_min using this to avoid function name confiction.}
\alias{dist_min}
\title{dist_min}
\usage{
dist_min(dt, q_cell, cell)
}
\arguments{
\item{sce}{object, a singcellexperiment object.}

\item{tocell}{string, CD8T.}
}
\value{
A data frame.
}
\description{
dist_min
}
\examples{
\dontrun{
# usage: calculate the distance to the nearest Epithelial_normal cell for each CD8T cell
sce <- readRDS("../05.anno/all_anno.rds")
coldata <- colData(sce) \%>\% as.data.table
res <- coldata[, dist_min(.SD, "CD8T", "Epithelial_normal"), by = sample_tiff_id]
}
}
