% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{.normalize_quantile}
\alias{.normalize_quantile}
\title{.normalize_quantile}
\usage{
.normalize_quantile(
  x,
  method = "qt_norm",
  range = c(-2.5, 2.5),
  quantile = 0.99
)
}
\arguments{
\item{x}{value vector}

\item{range}{zscore sd range.}
}
\description{
.normalize_quantile
}
