% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_function.R
\name{entropy_calc using this to avoid function name confiction.}
\alias{entropy_calc using this to avoid function name confiction.}
\alias{entropy_calc}
\title{entropy_calc}
\usage{
entropy_calc(sce, keycol = "cell_type10")
}
\arguments{
\item{sce}{object, a singcellexperiment object.}
}
\value{
A data frame.
}
\description{
entropy_calc
}
\examples{
\dontrun{
sce <- entropy_calc(sce)
sce <- entropy_calc(sce)
}
}
