% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_reducedims.R
\name{fetch_heatmap_dat using this to avoid function name confiction.}
\alias{fetch_heatmap_dat using this to avoid function name confiction.}
\alias{fetch_heatmap_dat}
\title{fetch_heatmap_dat}
\usage{
fetch_heatmap_dat(
  sce,
  assay = "log2counts_censored",
  cluster.id = "phenograph.id",
  scale = T,
  fil_markers = NULL,
  prefix = ""
)
}
\arguments{
\item{sce}{object, a singcellexperiment object.}

\item{assay}{string, "logcounts", "counts".}

\item{scale}{bool}
}
\value{
A list.
}
\description{
fetch_heatmap_dat
}
