% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_reducedims.R
\name{fetch_reducedims using this to avoid function name confiction.}
\alias{fetch_reducedims using this to avoid function name confiction.}
\alias{fetch_reducedims}
\title{fetch_reducedims}
\usage{
fetch_reducedims(sce, assay = "log2counts_censored", verbose = FALSE)
}
\arguments{
\item{sce}{object, a singcellexperiment object.}
}
\value{
A data frame.
}
\description{
fetch_reducedims
}
\examples{
\dontrun{
sce <- fetch_reducedims(sce)
sce <- fetch_reducedims(sce)
}
}
