% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cells.R
\name{filter_cells using this to avoid function name confiction.}
\alias{filter_cells using this to avoid function name confiction.}
\alias{filter_cells}
\title{filter_cells}
\usage{
filter_cells(
  sce,
  assay = "compcounts",
  cluster.id = NULL,
  phenograph.id = NULL,
  som.id = NULL,
  cell_id = NULL
)
}
\arguments{
\item{sce}{object, a singcellexperiment object.}
}
\value{
a singcellexperiment object
}
\description{
filter_cells
}
\examples{

sce <- filter_cells(sce, cluster.id = 1)
sce <- filter_cells(sce, cluster.id = c(1,2))
a <- filter_cells(sce, cluster.id = 35)
}
