% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{filter_cells_by_area using this to avoid function name confiction.}
\alias{filter_cells_by_area using this to avoid function name confiction.}
\alias{filter_cells_by_area}
\title{filter_cells_by_area}
\usage{
filter_cells_by_area(sce, sd_num = 4)
}
\arguments{
\item{sce}{object, a singcellexperiment object.}
}
\value{
a singcellexperiment object
}
\description{
Filter ion makers. We think most of markers will be kept. Therefore, we
              feed a makers vector will be filtered.
}
