% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_makers.R
\name{filter_makers using this to avoid function name confiction.}
\alias{filter_makers using this to avoid function name confiction.}
\alias{filter_makers}
\title{filter_makers}
\usage{
filter_makers(sce, makers = c("80ArAr", "120Sn"))
}
\arguments{
\item{sce}{object, a singcellexperiment object.}

\item{makers}{string vector, makers will be removed.}
}
\value{
a singcellexperiment object
}
\description{
Filter ion makers. We think most of markers will be kept. Therefore, we
              feed a makers vector will be filtered.
}
\examples{
\dontrun{
  sce <- filter_makers(sce, makers = 2000)
  sce <- filter_makers(sce, makers = cells)
}
}
