% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{filter_rare_marker_exp using this to avoid function name confiction.}
\alias{filter_rare_marker_exp using this to avoid function name confiction.}
\alias{filter_rare_marker_exp}
\title{filter_rare_marker_exp}
\usage{
filter_rare_marker_exp(sce, exps_markers_num = 5)
}
\arguments{
\item{sce}{object, a singcellexperiment object.}
}
\value{
a singcellexperiment object
}
\description{
Filter ion makers. We think most of markers will be kept. Therefore, we
              feed a makers vector will be filtered.
}
