% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_function.R
\name{getdiff_hyp using this to avoid function name confiction.}
\alias{getdiff_hyp using this to avoid function name confiction.}
\alias{getdiff_hyp}
\title{getdiff_hyp}
\usage{
getdiff_hyp(clinical_dat, keycol, feature_dat, method = "t")
}
\arguments{
\item{sce}{object, a singcellexperiment object.}
}
\value{
A data frame.
}
\description{
getdiff_hyp
}
\examples{
\dontrun{
sce <- getdiff_hyp(sce)
sce <- getdiff_hyp(sce)
}
}
