% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap_anno_combined}
\alias{heatmap_anno_combined}
\title{Visualization heatmap of data of sce}
\usage{
heatmap_anno_combined(
  sce,
  keycol = "cell_type10",
  nbhood_result,
  comm_clust,
  plots_positions
)
}
\arguments{
\item{sce}{A sce object}

\item{assay_name}{vector. Colors used in heatmap.}

\item{feature}{vector. markers to plot on the heatmap}

\item{scale}{character. Whether the values should be centered and scaled in either
the row direction or the column direction, or none. Corresponding values are
"row", "column" and "none"}

\item{col_fun}{logical. Whether columns should be clustered}

\item{...}{options to pass on to the \code{\link[ComplexHeatmap]{ComplexHeatmap}} function.}
}
\value{
ComplexHeatmap Heatmap figure
}
\description{
Visualization heatmap of data of sce
}
\examples{
\dontrun{
all_anno <- readr::read_rds("/cluster/home/jhuang/projects/hyperion/analysis/qzhang/human/steinbock/rds/combined/all_anno.rds")
pp <- c(0, 0.117, 0.25, 0.863,
         0.255, 0.001, 0.44, 0.98,
        0.705, 0.097, 0.3, 0.888,
        0.705, 0.097, 0.3, 0.888,
        0.85, 0.63, 0.02, 0.1)
pp <- matrix(pp, byrow = T, nrow = 5)
source("/cluster/home/yjliu_jh/projects/hyperion/code/plot/heatmap/main_heatmap_function.R")
heatmap_func(all_anno, plots_positions = pp)
}
}
