% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{imcAsinh_trans}
\alias{imcAsinh_trans}
\title{imcAsinh_trans}
\usage{
imcAsinh_trans(sce, assay = "compcounts")
}
\arguments{
\item{sce}{object, a singcellexperiment object}
}
\description{
imcAsinh_trans
}
\examples{
\dontrun{
sce <- imcAsinh_trans(sce)
assayNames(sce)
}

}
