% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{init_metadata using this to avoid function name confiction.}
\alias{init_metadata using this to avoid function name confiction.}
\alias{init_metadata}
\title{init_metadata}
\usage{
init_metadata(sce, sample_id, stype = "tumor")
}
\arguments{
\item{sce}{object, a singcellexperiment object.}
}
\value{
a singcellexperiment object
}
\description{
init_metadata
}
\examples{

sce <- init_metadata(sce, sample_id = "p155_pdac")
sce <- init_metadata(sce, sample_id = "p155_pdac", stype = "tumor")

}
