% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{interaction_heatmap using this to avoid function name confiction.}
\alias{interaction_heatmap using this to avoid function name confiction.}
\alias{interaction_heatmap}
\title{interaction_heatmap}
\usage{
interaction_heatmap(
  tiffs,
  res,
  sce,
  filename,
  cell_types = c("Bcell", "CD4T", "CD8T", "DC", "M_MDSC", "Monocyte",
    "Macrophage_CD163p", "Macrophage_HLADRp", "Macrophage_HLADRn_CD163n",
    "Macrophage_HLADRp_CD163p", "MMT_CD163p", "MMT_HLADRp", "MMT_HLADRn_CD163n",
    "MMT_HLADRp_CD163p", "mCAF", "CAF_col1p", "Vimentin_cell", "Epithelial_tumor",
    "Ki67_cell", "Epithelial_normal", "PSC", "Endothelial", "B7_H4_cell", "CCR6_cell")
)
}
\arguments{
\item{sce}{object, a singcellexperiment object.}
}
\value{
a singcellexperiment object
}
\description{
interaction_heatmap
}
