% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhuanglabHyperion-package.R
\docType{package}
\name{jhuanglabHyperion-package}
\alias{jhuanglabHyperion-package}
\alias{jhuanglabHyperion}
\title{Visualization and analyzation for Imaging Mass Cytometry™ (IMC™) with the Hyperion™ Imaging System data}
\description{
Functions and methods to visualize and analyze Imaging Mass Cytometry™ (IMC™) with the Hyperion™ Imaging
System data.
}
\details{
\tabular{ll}{ Package: \tab jhuanglabHyperion \cr Type: \tab Package \cr Version: \tab
1.0.0 \cr Date: \tab 2021-12-12 \cr License: \tab GPL-3.0 \cr }
While high-dimensional single-cell based Imaging Mass Cytometry data has
demonstrated increased applications in microenvironment composition and
solid tumor research, integrated analyzing workflow design for experimental
cytometry data has been challenging. Here, we present jhuanglabHyperion, an R package
designed for the analysis and interpretation of Imaging Mass Cytometry
data. We have applied jhuanglabHyperion to well designed compared-groups Imaging Mass Cytometry
Pancreatic ductal adenocarcinoma (PDAC) data to validate the usage and practical utility of its
computational modules. These use cases introduce jhuanglabHyperion as a reliable tool
for high-dimensional cytometry data workflow and reveal good performance
on trajectory reconstruction and pseudotime estimation.
}
\examples{

if (FALSE) {
## examples go here
## See vignette tutorials
vignette(package = "jhuanglabHyperion")
}
}
\author{
Maintainer: Jinyan Huang <huangjinyan@zju.edu.cn>
Authors: BMBDC of ZY
}
\keyword{package}
