% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{lgcl_transform}
\alias{lgcl_transform}
\title{lgcl_transform}
\usage{
lgcl_transform(sce, assay_name = "counts")
}
\arguments{
\item{sce}{object, a singcellexperiment object}
}
\description{
lgcl_transform
}
