% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{logcounts}
\alias{logcounts}
\alias{log2counts}
\title{logcounts}
\usage{
log2counts(sce, assay = "compcounts")
}
\arguments{
\item{sce}{object, a singcellexperiment object}
}
\description{
logcounts
}
