% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{lognormcounts}
\alias{lognormcounts}
\title{lognormcounts}
\usage{
lognormcounts(sce, assay = "compcounts", verbose = F)
}
\arguments{
\item{sce}{object, a singcellexperiment object}
}
\description{
lognormcounts
}
