% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{lognormcounts_cell}
\alias{lognormcounts_cell}
\title{lognormcounts_cell}
\usage{
lognormcounts_cell(sce, assay = "compcounts")
}
\arguments{
\item{sce}{object, a singcellexperiment object}
}
\description{
lognormcounts_cell
}
