% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rds.R, R/workflow.R
\name{make_rds}
\alias{make_rds}
\alias{workflow}
\title{make_rds}
\usage{
make_rds(
  stype = "normal",
  outputdir = "~/projects/hyperion/analysis/qzhang/human/steinbock/rds",
  dataset = "qzhang",
  overwrite = F,
  filter_makers = c("80ArAr", "120Sn", "127I", "134Xe", "138Ba", "DNA1", "DNA2",
    "Histone3", "208Pb"),
  glob = NULL
)

workflow(
  sce,
  assay = "logcounts_censored",
  outputdir = "~/projects/hyperion/analysis/qzhang/human/steinbock/rds/merge"
)
}
\arguments{
\item{stype}{a string c("paracancerous", "chemotherapy", "normal", "tumor", "puncture")}

\item{outputdir}{a string}

\item{assay}{a string c("logcounts_censored", "logcounts", "zscore")}
}
\description{
make_rds

make_rds
}
