% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaClustering.R
\name{map_to_singlecells}
\alias{map_to_singlecells}
\title{MetaClustering, elbow_test and tsne visualisation}
\usage{
map_to_singlecells(cc_metacluster, fSOM, view_tsne = T, seed = NULL)
}
\arguments{
\item{view_tsne}{TRUE/FALSE, determine whether tsne visualisation is performed, use TRUE as default.}

\item{seed, perplexity, max_iter}{the tsne parameters used in visualisation}

\item{indataframe}{event expression matrix or dataframe， containing one clustering channel}

\item{clustername}{string,the column name of clustering channel}

\item{metaClustering_method}{string,the method of metaclustering to use. could be set to "metaClustering_consensus",
"metaClustering_hclust","metaClustering_kmeans","metaClustering_PhenoGraph"}

\item{usecol}{numeric, the numeric columns id of indataframe, data in selected columns would be used in metaclustering.}

\item{elbow_test}{TRUE/FALSE, determine whether elbow_test will be carried out before formal metaclustering.}

\item{k_value}{numeric, the k value which is used in the metaclustering_method, equal to the number of generated clusters except for PhenoGraph.}
}
\description{
MetaClustering, elbow_test and tsne visualisation
}
