% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline_detect.R
\name{marker_IQR_delete remove outlier.}
\alias{marker_IQR_delete remove outlier.}
\alias{marker_IQR_delete}
\title{marker_IQR_delete}
\usage{
marker_IQR_delete(sce, marker = "CD45", iqr_num = 5)
}
\arguments{
\item{sce}{single cell experiment object.}
}
\description{
marker_IQR_delete
}
