% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline_detect.R
\name{marker_outlier_delete remove outlier.}
\alias{marker_outlier_delete remove outlier.}
\alias{marker_outlier_delete}
\title{marker_outlier_delete}
\usage{
marker_outlier_delete(sce, marker = "CD45", k = 20)
}
\arguments{
\item{sce}{single cell experiment object.}
}
\description{
http://r-statistics.co/Outlier-Treatment-With-R.html
    https://statsandr.com/blog/outliers-detection-in-r/
    https://rstudio-pubs-static.s3.amazonaws.com/300795_d5c82be1d7c548d881b78216e0b51c5d.html
    https://stackoverflow.com/questions/41462073/multivariate-outlier-detection-using-r-with-probability
    https://stackoverflow.com/questions/45289225/removing-multivariate-outliers-with-mvoutlier
}
