% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotaion.R
\name{merge_cell_type using this to avoid function name confiction.}
\alias{merge_cell_type using this to avoid function name confiction.}
\alias{merge_cell_type}
\title{merge_cell_type}
\usage{
merge_cell_type(sce, from = "cell_type", to = "cell_type10")
}
\arguments{
\item{sce}{data frame, a singcellexperiment object.}

\item{cell_anno}{a data frame.}
}
\description{
merge_cell_type
}
