% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cluster.R
\name{meta_clustering}
\alias{meta_clustering}
\title{meta_clustering}
\usage{
meta_clustering(
  object,
  assay = "log2counts_censored",
  secclustername = "FlowSOM",
  metaClustering_method = "metaClustering_PhenoGraph",
  k_value = 10,
  elbow_test = F,
  xdim = 40,
  ydim = 40,
  seed = 123
)
}
\arguments{
\item{object}{singlecellexperiment object;}

\item{assay}{string, posible value are: "counts" "imcAsinh" "simpleAsinh" "log2counts"}

\item{secclustername}{string character.}

\item{metaClustering_method}{string}

\item{k_value}{integer}

\item{elbow_test}{bool}

\item{xdim}{integer, this is for FlowSOM::SOM}

\item{ydim}{integer, this is for FlowSOM::SOM}

\item{seed}{integer, set seed.}
}
\description{
Uses a kd-tree to find the p number of
   This method is from xinlei cheng's package.
}
