% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{normalize_hyp}
\alias{normalize_hyp}
\title{normalize_hyp}
\usage{
normalize_hyp(
  sce,
  method = c("autoLgcl", "lognormcounts", "scater", "log2counts"),
  assay = "compcounts"
)
}
\arguments{
\item{sce}{object, a singcellexperiment object}
}
\description{
normalize_hyp
}
