% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{normalize_quantile}
\alias{normalize_quantile}
\title{normalize_hyp}
\usage{
normalize_quantile(obj, assay = "compcounts", ...)
}
\arguments{
\item{obj}{value vector}

\item{assay}{zscore sd range.}
}
\description{
normalize_hyp
}
