% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_function.R
\name{pairwise_comparison using this to avoid function name confiction.}
\alias{pairwise_comparison using this to avoid function name confiction.}
\alias{pairwise_comparison}
\title{pairwise_comparison}
\usage{
pairwise_comparison(feature_dat, class_info, method = "t")
}
\arguments{
\item{sce}{object, a singcellexperiment object.}
}
\value{
A data frame.
}
\description{
pairwise_comparison
}
\examples{
\dontrun{
sce <- pairwise_comparison(sce)
sce <- pairwise_comparison(sce)
}
}
