% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_dim.R
\name{pca_hyp}
\alias{pca_hyp}
\title{pca_hyp}
\usage{
pca_hyp(
  sce,
  assay = "log2counts_censored",
  center = TRUE,
  scale. = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sce}{object, a singcellexperiment object}
}
\description{
pca_hyp
}
