% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Visualization heatmap of data of sce}
\usage{
plot_heatmap(tiffs, res, sce, filenam)
}
\arguments{
\item{sce}{A sce object}

\item{assay_name}{vector. Colors used in heatmap.}

\item{feature}{vector. markers to plot on the heatmap}

\item{scale}{character. Whether the values should be centered and scaled in either
the row direction or the column direction, or none. Corresponding values are
"row", "column" and "none"}

\item{col_fun}{logical. Whether columns should be clustered}

\item{...}{options to pass on to the \code{\link[ComplexHeatmap]{ComplexHeatmap}} function.}
}
\value{
ComplexHeatmap Heatmap figure
}
\description{
Visualization heatmap of data of sce
}
