% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plotClusterHeatmap}
\alias{plotClusterHeatmap}
\title{Visualization heatmap of cluster data of sce}
\usage{
plotClusterHeatmap(
  object,
  color = colorRampPalette(c("blue", "white", "red"))(100),
  scale = "row",
  ...
)
}
\arguments{
\item{object}{A sce object}

\item{color}{vector. Colors used in heatmap.}

\item{scale}{character. Whether the values should be centered and scaled in either
the row direction or the column direction, or none. Corresponding values are
"row", "column" and "none"}

\item{...}{options to pass on to the \code{\link[pheatmap]{pheatmap}} function.}
}
\value{
ggplot2 figure
}
\description{
Visualization heatmap of cluster data of sce
}
\examples{

sce.file <- system.file("extdata/sce.rds", package = "jhuanglabHyperion")
sce <- readRDS(file = sce.file)

plotClusterHeatmap(sce)
plotClusterHeatmap(sce, color = colorRampPalette(c("purple","white","yellow"))(100))
plotClusterHeatmap(sce, cluster_row = FALSE)
plotClusterHeatmap(sce, cluster_row = FALSE, cluster_col = FALSE)



}
