% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Visualization heatmap of data of sce}
\usage{
plotHeatmap(
  object,
  markers = NULL,
  color = colorRampPalette(c("blue", "white", "red"))(100),
  scale = "row",
  downsize = 1000,
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  ...
)
}
\arguments{
\item{object}{A sce object}

\item{markers}{vector. markers to plot on the heatmap}

\item{color}{vector. Colors used in heatmap.}

\item{scale}{character. Whether the values should be centered and scaled in either
the row direction or the column direction, or none. Corresponding values are
"row", "column" and "none"}

\item{downsize}{numeric. Cells size used to plot heatmap}

\item{cluster_rows}{logical. Whether rows should be clustered}

\item{cluster_cols}{logical. Whether columns should be clustered}

\item{...}{options to pass on to the \code{\link[pheatmap]{pheatmap}} function.}
}
\value{
ggplot2 figure
}
\description{
Visualization heatmap of data of sce
}
\examples{

sce.file <- system.file("extdata/sce.rds", package = "jhuanglabHyperion")
sce <- readRDS(file = sce.file)

plotHeatmap(sce)
plotHeatmap(sce, cluster_rows = TRUE)
plotHeatmap(sce, cluster_rows = TRUE, clustering_method = "ward.D")
plotHeatmap(sce, cluster_rows = TRUE, cluster_cols = TRUE)


}
