% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2d.R
\name{plotPieCluster}
\alias{plotPieCluster}
\title{Visualization pie plot of cluster data of sce}
\usage{
plotPieCluster(
  object,
  item.use = c("PC_1", "PC_2"),
  cex.size = 1,
  size.by.cell.number = TRUE,
  main = "2D pie plot of sce",
  plot.theme = theme_bw()
)
}
\arguments{
\item{object}{A sce object}

\item{item.use}{character. Items use to 2D plot, axes x and y must be numeric.}

\item{cex.size}{numeric. Size of the dot}

\item{size.by.cell.number}{logical. Whether to show size of cell number.}

\item{main}{character. Title of the plot.}

\item{plot.theme}{themes from \code{ggplot2}}
}
\value{
ggplot2 figure
}
\description{
Visualization pie plot of cluster data of sce
}
\examples{

sce.file <- system.file("extdata/sce.rds", package = "imctoolkits")
sce <- readRDS(file = sce.file)

# Runs only have more than two stages
plotPieCluster(sce, cex.size = 0.5)

plotPieCluster(sce, item.use = c("PC_1", "PC_2"), cex.size = 0.5)
plotPieCluster(sce, item.use = c("PC_2", "PC_3"), cex.size = 0.5)

plotPieCluster(sce, item.use = c("tSNE_1", "tSNE_2"), cex.size = 20)

plotPieCluster(sce, item.use = c("DC_1", "DC_2"), cex.size = 0.5)

plotPieCluster(sce, item.use = c("UMAP_1", "UMAP_2"), cex.size = 1)
plotPieCluster(sce, item.use = c("UMAP_1", "UMAP_2"), cex.size = 1)


}
