% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_steinbock_img.R
\name{prepare_steinbock_imgcsv}
\alias{prepare_steinbock_imgcsv}
\title{prepare_steinbock_imgcsv}
\usage{
prepare_steinbock_imgcsv(
  dataset = "qzhang",
  stypes = c("paracancerous", "chemotherapy", "normal", "tumor", "puncture")
)
}
\description{
this is a quick an dirty function.
}
