% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_function.R
\name{quick_violin using this to avoid function name confiction.}
\alias{quick_violin using this to avoid function name confiction.}
\alias{quick_violin}
\title{quick_violin}
\usage{
quick_violin(data, x, y, z = NULL, ymin = -0.2, ymax = 0.9, step = 0.2, title)
}
\arguments{
\item{sce}{object, a singcellexperiment object.}
}
\value{
A data frame.
}
\description{
quick_violin
}
\examples{
\dontrun{
sce <- quick_violin(sce)
sce <- quick_violin(sce)
}
}
