% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_function.R
\name{quick_violin2 using this to avoid function name confiction.}
\alias{quick_violin2 using this to avoid function name confiction.}
\alias{quick_violin2}
\title{quick_violin2}
\usage{
quick_violin2(data, x, y = "sparse_score", z = "cell_type10")
}
\arguments{
\item{sce}{object, a singcellexperiment object.}
}
\value{
A data frame.
}
\description{
quick_violin2
}
\examples{
\dontrun{
sce <- quick_violin2(sce)
sce <- quick_violin2(sce)
}
}
