% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{range_scale}
\alias{range_scale}
\title{range_scale}
\usage{
range_scale(sce, assay = "log2counts_censored")
}
\arguments{
\item{sce}{object, a singcellexperiment object}
}
\description{
range_scale
}
