% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cluster.R
\name{run_clara}
\alias{run_clara}
\title{run_clara}
\usage{
run_clara(
  object,
  assay = "log2counts_censored",
  k = 25,
  metric = c("euclidean", "manhattan", "jaccard"),
  stand = FALSE,
  samples = 5,
  scale = TRUE,
  trace = 0,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{a sce object}

\item{k}{numeric. The number of clusters. It is required that
0 < k < n where n is the number of observations (i.e., n = nrow(x)).}

\item{metric}{character. string specifying the metric to be used for
calculating dissimilarities between observations.}

\item{stand}{logical. Indicating if the measurements in x are
standardized before calculating the dissimilarities.}

\item{samples}{numeric. Say N, the number of samples to be drawn from the dataset.
The default is N = 5,}

\item{scale}{logical. Whether to use scaled data in kmeans.}

\item{trace}{numberic. Indicating a trace level for diagnostic output during the algorithm}

\item{verbose}{logical. Whether to print calculation progress.}

\item{...}{Parameters passing to \code{\link[cluster]{clara}} function}
}
\value{
a sce object with clara.id in meta.data
}
\description{
Clustering a data matrix into k clusters
}
\examples{

sce.file <- system.file("extdata/sce.rds", package = "sceoTree")
sce <- readRDS(file = sce.file)

sce <- run_clara(sce, k = 25, verbose = TRUE)


}
\seealso{
\code{\link[cluster]{clara}}
}
