% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cluster.R
\name{run_cluster}
\alias{run_cluster}
\title{Specific Clustering Method Toolkits
Adopt these methods from sceoTree.}
\usage{
run_cluster(
  object,
  assay = "log2counts_censored",
  cluster.method = c("fastpg", "som", "kmeans", "clara", "phenograph", "hclust",
    "mclust"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{a sce object}

\item{cluster.method}{character. Four clustering method are provided: som, clara, kmeans and phenograph.
Clustering method "hclust" and "mclust" are not recommended because of long computing time.}

\item{verbose}{logic. Whether to print calculation progress.}

\item{...}{options to pass on to the clustering functions.}
}
\value{
A sce object with cluster
}
\description{
Compute a specific clustering using the combined flow
   sceometry data. "som" \code{\link[FlowSOM]{SOM}}, "hclust" \code{\link[stats]{hclust}},
   "clara" \code{\link[cluster]{clara}}, "phenograph", "kmeans" \code{\link[stats]{kmeans}} are
   provided.
}
\examples{

sce.file <- system.file("extdata/sce.rds", package = "sceoTree")
sce <- readRDS(file = sce.file)

# After building an sce object
# Set random seed to make results reproducible

set.seed(1)
sce <- run_cluster(sce, cluster.method = "som", xdim = 3, ydim = 3, verbose = TRUE)

# K-means clustering
sce <- run_cluster(sce, cluster.method = "kmeans", k = 9, verbose = TRUE)

# Clara clustering
sce <- run_cluster(sce, cluster.method = "clara", k = 9, verbose = TRUE)

# phenoGraph clustering
sce <- run_cluster(sce, cluster.method = "phenograph", verbose = TRUE)

# hclust clustering
# not recommended for large cell size
sce <- run_cluster(sce, cluster.method = "hclust", k = 9, verbose = TRUE)

# mclust clustering
# not recommended for large cell size
sce <- run_cluster(sce, cluster.method = "mclust", verbose = TRUE)


}
\seealso{
\code{\link[FlowSOM]{SOM}}, \code{\link[stats]{hclust}},
   \code{\link[cluster]{clara}}, \code{\link[stats]{kmeans}}.
   You can use \code{run_SOM}, \code{run_clara},
   \code{runPhenotype}, \code{run_kmeans}, \code{run_mclust} and
   \code{run_hclust} to run clustering respectively.
}
