% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cluster.R
\name{run_fastpg}
\alias{run_fastpg}
\title{run_fastpg clustering}
\usage{
run_fastpg(object, assay = "log2counts_censored", k = 30, verbose = FALSE, ...)
}
\arguments{
\item{object}{a sce object.}

\item{k}{numeric. Number of nearest neighbours, default is 30.}

\item{...}{Parameters passing to \code{igraph} function}
}
\value{
A sce object with cluster
}
\description{
A simple R implementation of the FastPG
   [FastPG](https://github.com/sararselitsky/FastPG) algorithm
}
\examples{

sce.file <- system.file("extdata/sce.rds", package = "sceoTree")
sce <- readRDS(file = sce.file)

sce <- run_phenograph(sce, knn = 30, verbose = TRUE)


}
