% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cluster.R
\name{run_mclust}
\alias{run_mclust}
\title{run_mclust}
\usage{
run_mclust(
  object,
  assay = "log2counts_censored",
  scale = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{a sce object}

\item{scale}{logical. Whether to use scaled data in Mclust.}

\item{verbose}{logical. Whether to print calculation progress.}

\item{...}{Parameters passing to \code{\link[mclust]{Mclust}} function}
}
\value{
a sce object with mclust.id in meta.data
}
\description{
Model-based clustering based on parameterized finite Gaussian mixture models.
   This function is based on \code{\link[mclust]{Mclust}}.
}
\examples{

sce.file <- system.file("extdata/sce.rds", package = "sceoTree")
sce <- readRDS(file = sce.file)

sce <- run_mclust(sce, verbose = TRUE)


}
\seealso{
\code{\link[mclust]{Mclust}}
}
