% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cluster.R
\name{run_phenograph}
\alias{run_phenograph}
\title{RphenoGraph clustering}
\usage{
run_phenograph(
  object,
  assay = "log2counts_censored",
  k = 30,
  scale = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{a sce object.}

\item{scale}{logical. Whether to scale the expression matrix}

\item{verbose}{logical. Whether to print calculation progress.}

\item{...}{Parameters passing to \code{igraph} function}

\item{knn}{numeric. Number of nearest neighbours, default is 30.}
}
\value{
A sce object with cluster
}
\description{
A simple R implementation of the phenograph
   [PhenoGraph](http://www.cell.com/cell/abstract/S0092-8674(15)00637-6) algorithm,
   which is a clustering method designed for high-dimensional single-cell
   data analysis. It works by creating a graph ("network") representing
   phenotypic similarities between cells by calculating the Jaccard
   coefficient between nearest-neighbor sets, and then identifying communities
   using the well known [Louvain method](https://sites.google.com/site/findcommunities/)
   in this graph.
}
\examples{

sce.file <- system.file("extdata/sce.rds", package = "sceoTree")
sce <- readRDS(file = sce.file)

sce <- run_phenograph(sce, knn = 30, verbose = TRUE)


}
