% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{sd_censor}
\alias{sd_censor}
\alias{sd_censor_max}
\title{sd_censor}
\usage{
sd_censor(x, range)

sd_censor_max(x, range)
}
\arguments{
\item{x}{value vector}

\item{range}{zscore sd range.}
}
\description{
sd_censor

sd_censor
}
