% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{simpleAsinh_trans}
\alias{simpleAsinh_trans}
\title{simpleAsinh_trans}
\usage{
simpleAsinh_trans(sce, assay = "compcounts")
}
\arguments{
\item{sce}{object, a singcellexperiment object}
}
\description{
simpleAsinh_trans
}
