% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{simpleAsinh_Zero_Max}
\alias{simpleAsinh_Zero_Max}
\title{simpleAsinh_Zero_Max}
\usage{
simpleAsinh_Zero_Max(value, cofactor = 5, b = 0)
}
\arguments{
\item{value}{object, a singcellexperiment object}
}
\description{
simpleAsinh_Zero_Max
}
