% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steinbock.R
\name{steinbock}
\alias{steinbock}
\title{steinbock
conda activate steinbock
Wrapper for steinbock python package, steinbock need to be installed.}
\usage{
steinbock(stype, sample_id, panel_fn, dmax = 30, analysis_dir)
}
\arguments{
\item{stype}{character string, either c("normal", "paracancerous", "puncture", "tumor")}

\item{sample_id}{character string, for example: p155_pdac}

\item{panel_fn}{character string, path of panel file.
for example: panel_fn <- glue("~/projects/{project}/data/{dataset}/{species}/hyperion/panel.csv")}

\item{dmax}{integer, default 30}

\item{analysis_dir}{character string, a path where the analysis output.}
}
\description{
steinbock
conda activate steinbock
Wrapper for steinbock python package, steinbock need to be installed.
}
\examples{
\dontrun{
conda activate steinbock
steinbock(stype, sample_id, panel_fn, dmax = 30, analysis_dir)
We will setup run steinbock with singularity or docker.
}
}
