% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_hyp.R
\name{subset_by_cell_ids using this to avoid function name confiction.}
\alias{subset_by_cell_ids using this to avoid function name confiction.}
\alias{subset_by_cell_ids}
\title{subset_by_cell_ids}
\usage{
subset_by_cell_ids(sce, cell_ids, include = TRUE)
}
\arguments{
\item{sce}{object, a singcellexperiment object.}

\item{cell_ids}{integer vector, c(45, 48, 52, 7).}

\item{include}{in or not in.}
}
\value{
a singcellexperiment object
}
\description{
subset_by_cell_ids
}
