% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_hyp.R
\name{subset_by_cluster_id using this to avoid function name confiction.}
\alias{subset_by_cluster_id using this to avoid function name confiction.}
\alias{subset_by_cluster_id}
\title{subset_by_cluster_id}
\usage{
subset_by_cluster_id(sce, fastpg, cluster_ids = c(45, 48, 52, 7))
}
\arguments{
\item{sce}{object, a singcellexperiment object.}

\item{fastpg}{data frame, a singcellexperiment object.}

\item{cluster_ids}{integer vector, c(45, 48, 52, 7).}
}
\value{
a singcellexperiment object
}
\description{
subset_by_cluster_id
}
