% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_hyp.R
\name{subset_by_feature using this to avoid function name confiction.}
\alias{subset_by_feature using this to avoid function name confiction.}
\alias{subset_by_feature}
\title{subset_by_feature}
\usage{
subset_by_feature(
  sce,
  feature_name = "cell_type",
  features = c("Unknown"),
  include = TRUE
)
}
\arguments{
\item{sce}{object, a singcellexperiment object.}

\item{features}{integer vector, c(45, 48, 52, 7).}

\item{feture_name}{data frame, a singcellexperiment object.}
}
\value{
a singcellexperiment object
}
\description{
subset_by_feature
}
