% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_hyp.R
\name{subset_by_marker_heatmap_exp using this to avoid function name confiction.}
\alias{subset_by_marker_heatmap_exp using this to avoid function name confiction.}
\alias{subset_by_marker_heatmap_exp}
\title{subset_by_marker_heatmap_exp}
\usage{
subset_by_marker_heatmap_exp(
  sce,
  marker = "CD45",
  fastpg,
  dat,
  cutoff,
  direction = "up"
)
}
\arguments{
\item{sce}{object, a singcellexperiment object.}

\item{marker}{string, "CD45".}

\item{fastpg}{data frame, a singcellexperiment object.}

\item{dat}{data frame, for heatmap plot.}

\item{cutoff}{double, a singcellexperiment object.}

\item{direction}{string, up or down.}
}
\value{
a singcellexperiment object
}
\description{
subset_by_marker_heatmap_exp
}
\examples{

sce <- subset_by_marker_heatmap_exp(sce, cell_number = 2000)
sce <- subset_by_marker_heatmap_exp(sce, cells = cells)

}
