% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_hyp.R
\name{subset_hyp using this to avoid function name confiction.}
\alias{subset_hyp using this to avoid function name confiction.}
\alias{subset_hyp}
\title{subset_hyp}
\usage{
subset_hyp(sce, cell_number = 3e+05, force = TRUE)
}
\arguments{
\item{sce}{object, a singcellexperiment object.}
}
\value{
a singcellexperiment object
}
\description{
subset_hyp
}
\examples{

sce <- subset_hyp(sce, cell_number = 2000)
sce <- subset_hyp(sce, cells = cells)

}
