% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_function.R
\name{temp_plot using this to avoid function name confiction.}
\alias{temp_plot using this to avoid function name confiction.}
\alias{temp_plot}
\title{temp_plot}
\usage{
temp_plot(data, list, u, group_var)
}
\arguments{
\item{sce}{object, a singcellexperiment object.}
}
\value{
A data frame.
}
\description{
temp_plot
}
\examples{
\dontrun{
sce <- temp_plot(sce)
sce <- temp_plot(sce)
}
}
