% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_dim.R
\name{tsne_hyp using this to avoid function name confiction.}
\alias{tsne_hyp using this to avoid function name confiction.}
\alias{tsne_hyp}
\title{tsne_hyp}
\usage{
tsne_hyp(sce, assay = "log2counts_censored", dims = 40)
}
\arguments{
\item{sce}{object, a singcellexperiment object. For more details, see \code{\link[Rtsne]{Rtsne}}.}
}
\value{
a singcellexperiment object
}
\description{
tsne_hyp
}
\examples{

sce <- normalize_hyp(sce)
sce <- tsne_hyp(sce)

}
\references{
Maaten, L. Van Der, 2014. Accelerating t-SNE using Tree-Based
   Algorithms. Journal of Machine Learning Research, 15, p.3221-3245.

   van der Maaten, L.J.P. & Hinton, G.E., 2008. Visualizing High-Dimensional
   Data Using t-SNE. Journal of Machine Learning Research, 9, pp.2579-2605.
}
\seealso{
\code{\link[Rtsne]{Rtsne}}
}
