% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_dim.R
\name{umap_hyp}
\alias{umap_hyp}
\title{umap_hyp}
\usage{
umap_hyp(sce, assay = "log2counts_censored", dims = 40, umap.method = "uwot")
}
\arguments{
\item{sce}{object, a singcellexperiment object}
}
\description{
umap_hyp
}
