% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotaion.R
\name{update_feature using this to avoid function name confiction.}
\alias{update_feature using this to avoid function name confiction.}
\alias{update_feature}
\title{update_feature}
\usage{
update_feature(sce, cell_anno, feature = "cell_type")
}
\arguments{
\item{sce}{data frame, a singcellexperiment object.}

\item{cell_anno}{a data frame.}
}
\description{
update_feature
}
