% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{update_metadata using this to avoid function name confiction.}
\alias{update_metadata using this to avoid function name confiction.}
\alias{update_metadata}
\title{update_metadata}
\usage{
update_metadata(
  sce,
  downsample = NULL,
  cell_type = NULL,
  sample_id = NULL,
  stype = NULL,
  df = NULL
)
}
\arguments{
\item{sce}{object, a singcellexperiment object.}
}
\value{
a singcellexperiment object
}
\description{
update_metadata
}
\examples{

sce <- update_metadata(sce)
sce <- update_metadata(sce, sample_id = "p155_pdac", stype = "tumor")

}
