% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{update_metadata_info using this to avoid function name confiction.}
\alias{update_metadata_info using this to avoid function name confiction.}
\alias{update_metadata_info}
\title{update_metadata_info}
\usage{
update_metadata_info(sce, sample_id = NULL, stype = NULL)
}
\arguments{
\item{sce}{object, a singcellexperiment object.}
}
\value{
a singcellexperiment object
}
\description{
update_metadata_info
}
\examples{

sce <- update_metadata_info(sce)
sce <- update_metadata_info(sce, sample_id = "p155_pdac", stype = "tumor")

}
