% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotaion.R
\name{write_anno_by_cid using this to avoid function name confiction.}
\alias{write_anno_by_cid using this to avoid function name confiction.}
\alias{write_anno_by_cid}
\title{write_anno_by_cid}
\usage{
write_anno_by_cid(fastpg, cluster_anno, fn_out)
}
\arguments{
\item{fastpg}{data frame, a singcellexperiment object.}

\item{cluster_anno}{a data frame.}

\item{fn_out}{string}
}
\description{
write_anno_by_cid
}
