% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotaion.R
\name{write_anno_sce_by_cid using this to avoid function name confiction.}
\alias{write_anno_sce_by_cid using this to avoid function name confiction.}
\alias{write_anno_sce_by_cid}
\title{write_anno_sce_by_cid}
\usage{
write_anno_sce_by_cid(
  sce,
  cluster_anno,
  cluster_id = "fastpg.id",
  fn_out,
  unknow_include = F
)
}
\arguments{
\item{cluster_anno}{a data frame.}

\item{cluster_id}{"fastpg.id"}

\item{fn_out}{string}

\item{fastpg}{data frame, a singcellexperiment object.}
}
\description{
write_anno_sce_by_cid
}
\examples{
\dontrun{
 write_anno_sce_by_cid(sce, cluster_anno, fn_out = fn_out)
}
write_anno_sce_by_cid(sce, cluster_anno, fn_out = fn_out)
}
