% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{x using this to avoid function name confiction.}
\alias{x using this to avoid function name confiction.}
\alias{fastercheck}
\title{fastercheck}
\usage{
fastercheck(x, matrix)
}
\arguments{
\item{matrix}{object, a singcellexperiment object.}
}
\value{
a singcellexperiment object
}
\description{
not export
}
