% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{zscore_censor}
\alias{zscore_censor}
\title{zscore_censor}
\usage{
zscore_censor(obj, assay = "compcounts", range = c(-2.5, 2.5))
}
\arguments{
\item{range}{zscore sd range.}

\item{sce}{object, a singcellexperiment object.}
}
\description{
Data transformation and normalization
  The presented data were not transformed, and all analyses were
  based on raw IMC measurements. Single-cell marker expressions are
  summarized by mean pixel values for each channel. The single-cell
  data were censored at the 99th percentile to remove outliers, and
  z-scored cluster means were visualized in heat maps. For t-SNE and
  PhenoGraph the data were normalized to the 99th percentile, as is
  suggested for these algorithms23,31. To visualize the number of cells
  per image or patient and for survival modelling, the cell counts were
  normalized by the image area (total number of pixels) and displayed
  as cell density. For Cox proportional hazards survival modelling, these
  densities were multiplied by a factor of 107 to yield values larger than 1
  and then log-transformed. ref: https://doi.org/10.1038/s41586-019-1876-x
}
