% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rds.R
\name{make_rds}
\alias{make_rds}
\title{make_rds}
\usage{
make_rds(
  stype = "normal",
  outputdir = "~/projects/hyperion/analysis/qzhang/human/steinbock/rds",
  dataset = "qzhang",
  overwrite = F,
  filter_makers = c("80ArAr", "120Sn", "127I", "134Xe", "138Ba", "DNA1", "DNA2",
    "Histone3", "208Pb"),
  glob = NULL
)
}
\arguments{
\item{stype}{a string c("paracancerous", "chemotherapy", "normal", "tumor", "puncture")}

\item{outputdir}{a string}
}
\description{
make_rds
}
